+++ tests/Integration/Schema/Directives/MorphOneDirectiveTest.php
@@ -10,7 +10,7 @@ use Tests\Utils\Models\User;
 class MorphOneDirectiveTest extends DBTestCase
 {
     /**
-     * Auth user.
+     * The authenticated user.
      *
      * @var \Tests\Utils\Models\User
      */
